requirements.txt
Pillow==7.1.1
beautifulsoup4==4.9.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
SQLAlchemy==1.3.15
WTForms==2.2.1
Werkzeug==1.0.0
alembic==1.4.2
blinker==1.4
boto3==1.12.31
botocore==1.15.31
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
dnspython==1.16.0
docutils==0.15.2
dominate==2.5.1
email-validator==1.0.5
flask-googlemaps==0.4.0
idna==2.9
itsdangerous==1.1.0
jmespath==0.9.5
pip==19.0.3
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
s3transfer==0.3.3
setuptools==40.8.0
six==1.14.0
urllib3==1.25.8
visitor==0.1.3
pyOpenSSL==19.1.0
Flask-Moment==0.9.0
flask-paginate==0.5.5
geojson==2.5.0
mapbox==0.18.0
elasticsearch==7.6.0
Flask-Cors==3.0.8
indeed==0.0.4
Flask-SSLify==0.1.5
flask-talisman==0.7.0
Flask-CDN==1.5.3
celery==4.4.6
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
google-auth-oauthlib==0.4.1
Flask-Execute==0.1.4
supersaas-api-client==1.2.2